chore: refactor into multiple files
