+++ recorder.py
@@ -1,0 +1,45 @@
+import pyaudio
+import wave
+import threading
+
+
+class AudioRecorder:
+    def __init__(self):
+        self.audio = pyaudio.PyAudio()
+        self.stream = None
+        self.frames = []
+        self.recording = False
+
+    def start_recording(self):
+        self.frames = []
+        self.stream = self.audio.open(
+            format=pyaudio.paInt16,
+            channels=1,
+            rate=44100,
+            input=True,
+            frames_per_buffer=1024,
+        )
+        self.recording = True
+        self.thread = threading.Thread(target=self.record)
+        self.thread.start()
+
+    def record(self):
+        while self.recording:
+            data = self.stream.read(1024)
+            self.frames.append(data)
+
+    def stop_recording(self):
+        self.recording = False
+        self.thread.join()
+        self.stream.stop_stream()
+        self.stream.close()
+        self.save_recording()
+
+    def save_recording(self):
+        self.filepath = "output.wav"
+        wf = wave.open(self.filepath, "wb")
+        wf.setnchannels(1)
+        wf.setsampwidth(self.audio.get_sample_size(pyaudio.paInt16))
+        wf.setframerate(44100)
+        wf.writeframes(b"".join(self.frames))
+        wf.close()
